Convert all 214 VO again but at 10512Hz
While EA2 definitely used 11025Hz, DTT sounded slightly different, so after reading about Bregalad's GBA sound doc, I realized that the pitch value defined before each sample, can reveal the sample rate by dividing the value by 1024, thus DTT uses 10512 Hz for all VO.

If a user replaces a sound in the importer, sounds should still be encoded at 11025Hz because the pitch value will be hardcoded to that in my program, it's also easier to select in Audacity.

If I can find some disadvantage I'll switch to 10512 but so far it seems like the only difference is it takes more space, but with the possibility of sounding better.