VO intros randomly selected
This is an optional hack that allows a character's intro voice to not collide with the other by only running the code once.

Ponygon and Karudio have double voice calls so they'll still have some conflicts.

The only other option is to make all the intros into new sound groups so that they can all play at the same time like in most fighting games, will probably try that later.