MORE voices done
Arth now has his VO filtered through openVINO to achieve a much cleaner result.

Victoream's "FUKKATSU" line is used in episode 89 a few times, the English dub translated the first one as "I'm baaaaack!" and the second as "I'm rejuvenated!" so I decided to pick a different clip for it where he yells "Come on!" which seems to fit well enough.