Better system for item name GFX
Now the raw data is mixed with a header that my importer will read to get the position on the ROM to patch the data in.

This makes it so that all item names are quickly checked from the importer using a single loop.

Added titles for items 1 thru 7 in English and Spanish.