Updated font with an alternate í
The accented "i" is problematic because it needs 3 pixels, and the VWF only works if the width is divisible by 2, so if you make this character 4 pixels wide it leaves a lot of empty space on the screen.

This alternate í is shifted to the right by 1px, making it look centered on the empty space.

The encoder has been updated to use this alternate character automatically if the following character is a->Z.

Furhermore the apostrophe character has been edited to use only 1 pixel(+space) so it's a 2px character now, allowing text to look nicer on screen.

The encoder has been updated to automatically handle the inverted/initial quotes, this character was already in the game.